moved these out of main
