+++ allstar_processing.py
