+++ modeling_snubs.py
